Full support of particles loading for all classes
All CCQuadParticle based classes have been supported with the
"particleWithFile" method in order to load particles from PLIST files
(see Particles Designer for MAC).

Additional changes have been made to CCTextureAtlas for supporting the
Java BufferUtils instead of the GDX which is either missing or not
compatible with the Windows version of Eclipse.